create component for each highscore
