+++ client/src/components/ScoreTabs/ScoreTabs.jsx
@@ -1,15 +1,10 @@
 import React from 'react'
-import { MDBTabPane, MDBBadge, MDBIcon, MDBAnimation, MDBTooltip } from "mdbreact";
-import { parseSongPicUrl } from '../../helper/parser'
-import moment from 'moment'
-import DiffTags from '../DiffTag'
-import HighscoreTable from './HighscoreTable'
+import { MDBTabPane } from "mdbreact";
+import OneHighScore from './OneHighScore'
 import '../../styles/ScoreTabs.scss'
 
 export default function ScoreTabs(props) {
     const { tabId, data, size } = props;
-
-    const logid = (scoreId) => console.log("scoreId: !", scoreId)
 
     if(data.length === 0) return null
     return (
@@ -27,39 +22,7 @@ export default function ScoreTabs(props) {
                                 </tr>
                             </thead>
                             <tbody>
-                                {data.map((data, i) => {
-                                    return(
-                                        <tr key={i} className={data.playedByHive ? "played-by-hive" : ""}>
-                                            <td><MDBBadge color="pink">{data.rank}</MDBBadge></td>
-                                            <td>
-                                                <div className="song-data">
-                                                    <DiffTags diff={data.difficulty} />
-                                                    <img src={parseSongPicUrl(data.songHash)} alt="Cover"/>
-                                                    {data.playedByHive && 
-                                                        <MDBTooltip domElement clickable tag="span" placement="top">
-                                                            <span><MDBAnimation type="pulse" infinite><MDBIcon far icon="handshake" /></MDBAnimation></span>
-                                                            <div className="also-played-by">
-                                                                <table>
-                                                                    <thead>
-                                                                        <th>Bee</th>
-                                                                        <th>Score</th>
-                                                                        <th>My Score</th>
-                                                                    </thead>
-                                                                    <tbody id="bees">
-                                                                        {data.playedBy.map((bee, i) => <HighscoreTable bee={bee} i={i} />)}
-                                                                    </tbody>
-                                                                </table>
-                                                                {/* {data.playedBy.map((bee, i) => <div id="bees" key={i} ><BeeTag bee={bee}/></div>)} */}
-                                                            </div>
-                                                        </MDBTooltip>}
-                                                    <MDBBadge onClick={() => logid(data.songHash)} color="dark">{data.songAuthorName} - {data.songName}<span className="greyed-out"> by {data.levelAuthorName}</span></MDBBadge>
-                                                </div>
-                                            </td>
-                                            <td><MDBBadge color="orange">{data.score}</MDBBadge></td>
-                                            <td className="time"><b className="card-link blue-text">{moment(data.timeSet).format('lll')}</b></td>
-                                        </tr>
-                                    )})
-                                }
+                                {data.map((highscore, i) => <OneHighScore highscore={highscore} key={i} />)}
                             </tbody>
                         </table>
                     </div>
